Fix WSDL files location
The WSDL and respective XSD and XJC files were located in the `src` directory.
However the `src` directory is only for Java source files. As such, the
resulting JAR would have all files in the root directory instead of `wsdl`,
`xsd` and `jaxws` directories. The WSDL would not have valid schema locations
and some web services frameworks, such as Apache CXF, will not allow to use the
WSDL client as is. To fix this issue these files were moved to
`src/main/resources`.